move feedback container to component
